[Refactor] #1701 quota evidence 검증 공용화
