+++ tools/datapack/build-admin-review-record.mjs
@@ -19,10 +19,9 @@ import {
   requiredString,
   sortJson,
 } from "./lib/ledger-admission-cli.mjs";
+import { validateQuotaEvidence } from "./lib/quota-evidence.mjs";
 
 const root = path.resolve(import.meta.dirname, "../..");
-const requiredQuotaEvidenceKeys = ["defaultDailyLimit", "portal", "productionUseAllowed", "unlockStatus"];
-const optionalQuotaEvidenceKeys = ["documentedMonthlyLimit"];
 
 async function main() {
   const args = parseArgs(process.argv.slice(2));
@@ -117,38 +116,6 @@ async function readProductionSource(args, sourceId, quotaEvidence, sampleEvidenc
     }
   }
   return productionSource;
-}
-
-function validateQuotaEvidence(quotaEvidence, label) {
-  if (!quotaEvidence || typeof quotaEvidence !== "object" || Array.isArray(quotaEvidence)) {
-    throw new TypeError(`${label} must be an object`);
-  }
-  const keys = Object.keys(quotaEvidence);
-  const allowedKeys = new Set([...requiredQuotaEvidenceKeys, ...optionalQuotaEvidenceKeys]);
-  if (!requiredQuotaEvidenceKeys.every((key) => keys.includes(key)) || keys.some((key) => !allowedKeys.has(key))) {
-    throw new Error(`${label} must include ${requiredQuotaEvidenceKeys.join(", ")} and only optional ${optionalQuotaEvidenceKeys.join(", ")}`);
-  }
-  requiredString(quotaEvidence.portal, `${label}.portal`);
-  if (
-    quotaEvidence.defaultDailyLimit !== null &&
-    quotaEvidence.defaultDailyLimit !== "unlimited" &&
-    (!Number.isInteger(quotaEvidence.defaultDailyLimit) || quotaEvidence.defaultDailyLimit < 0)
-  ) {
-    throw new Error(`${label}.defaultDailyLimit must be null, a non-negative integer, or unlimited`);
-  }
-  if (
-    "documentedMonthlyLimit" in quotaEvidence &&
-    (!Number.isInteger(quotaEvidence.documentedMonthlyLimit) || quotaEvidence.documentedMonthlyLimit <= 0)
-  ) {
-    throw new Error(`${label}.documentedMonthlyLimit must be a positive integer`);
-  }
-  if (quotaEvidence.defaultDailyLimit === null && !("documentedMonthlyLimit" in quotaEvidence)) {
-    throw new Error(`${label}.defaultDailyLimit null requires documentedMonthlyLimit`);
-  }
-  requiredString(quotaEvidence.unlockStatus, `${label}.unlockStatus`);
-  if (typeof quotaEvidence.productionUseAllowed !== "boolean") {
-    throw new TypeError(`${label}.productionUseAllowed must be a boolean`);
-  }
 }
 
 function assertSha256(value, label) {
