+++ tools/datapack/run-source-admission-pipeline.mjs
@@ -6,11 +6,10 @@ import path from "node:path";
 import { pathToFileURL } from "node:url";
 import { promisify } from "node:util";
 import { sortJson, compareStrings } from "./lib/ledger-admission-cli.mjs";
+import { validateQuotaEvidence } from "./lib/quota-evidence.mjs";
 
 const execFileAsync = promisify(execFile);
 const root = path.resolve(import.meta.dirname, "../..");
-const requiredQuotaEvidenceKeys = ["defaultDailyLimit", "portal", "productionUseAllowed", "unlockStatus"];
-const optionalQuotaEvidenceKeys = ["documentedMonthlyLimit"];
 
 async function main() {
   const startedAt = Date.now();
@@ -218,38 +217,6 @@ function validateAdminReview({ adminReview, candidateId, sample, snapshot, args 
     }
   }
   return sha256(JSON.stringify(sortJson(adminReview)));
-}
-
-function validateQuotaEvidence(quotaEvidence, label) {
-  if (!quotaEvidence || typeof quotaEvidence !== "object" || Array.isArray(quotaEvidence)) {
-    throw new Error(`${label} must be an object`);
-  }
-  const keys = Object.keys(quotaEvidence);
-  const allowedKeys = new Set([...requiredQuotaEvidenceKeys, ...optionalQuotaEvidenceKeys]);
-  if (!requiredQuotaEvidenceKeys.every((key) => keys.includes(key)) || keys.some((key) => !allowedKeys.has(key))) {
-    throw new Error(`${label} must include ${requiredQuotaEvidenceKeys.join(", ")} and only optional ${optionalQuotaEvidenceKeys.join(", ")}`);
-  }
-  requiredText(quotaEvidence.portal, `${label}.portal`);
-  if (
-    quotaEvidence.defaultDailyLimit !== null &&
-    quotaEvidence.defaultDailyLimit !== "unlimited" &&
-    (!Number.isInteger(quotaEvidence.defaultDailyLimit) || quotaEvidence.defaultDailyLimit < 0)
-  ) {
-    throw new Error(`${label}.defaultDailyLimit must be null, a non-negative integer, or unlimited`);
-  }
-  if (
-    "documentedMonthlyLimit" in quotaEvidence &&
-    (!Number.isInteger(quotaEvidence.documentedMonthlyLimit) || quotaEvidence.documentedMonthlyLimit <= 0)
-  ) {
-    throw new Error(`${label}.documentedMonthlyLimit must be a positive integer`);
-  }
-  if (quotaEvidence.defaultDailyLimit === null && !("documentedMonthlyLimit" in quotaEvidence)) {
-    throw new Error(`${label}.defaultDailyLimit null requires documentedMonthlyLimit`);
-  }
-  requiredText(quotaEvidence.unlockStatus, `${label}.unlockStatus`);
-  if (typeof quotaEvidence.productionUseAllowed !== "boolean") {
-    throw new Error(`${label}.productionUseAllowed must be a boolean`);
-  }
 }
 
 function admitSource({ inventory, productionSource }) {
